ohore: Move dpa_monitor out of crates/api
