+++ crates/api-core/src/cfg/file.rs
@@ -1995,7 +1995,7 @@ fn default_mqtt_broker_port() -> u16 {
     1884
 }
 
-pub use carbide_dpa_manager::cfg::{DpaConfig, MqttAuthConfig, MqttAuthMode};
+pub use carbide_dpa_manager::config::{DpaConfig, MqttAuthConfig, MqttAuthMode};
 
 /// DSX Exchange Event Bus configuration for publishing state change events via MQTT 3.1.1.
 ///
